git-svn-id: svn+ssh://ssh.dna.bio.keio.ac.jp/home/slab/satoken/svnroot/stem_kernel/trunk@198 0ffabbc1-8d0a-0410-9eb0-baa645350600

+++ stem_kernel_lite/stem_kernel.cpp
@@ -109,7 +109,8 @@ operator()(const Data& xx, const Data& yy) const
       // MATCH
       K1[j] = G1[j] = 0.0;
       if (!x[i].empty() && !y[j].empty() &&
-	  (len_band_==0 || abs(x[i].length()-y[j].length())<=len_band_) ) {
+	  (len_band_==0 ||
+	   static_cast<uint>(abs(x[i].length()-y[j].length()))<=len_band_) ) {
 	// score for this node
 	value_type v_s = st_.node_score(xx, yy, i, j);
 	for (ix=x[i].begin(); ix!=x[i].end(); ++ix) {
